requirements.txt
ultralytics==8.2.35
numpy==1.26.4
Pillow==10.3.0
opencv-python==4.9.0.80
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
gradio==4.19.0